タイトルシーンの修正
タイトルからステージ１に移動するボタンを追加
タイトルからステージセレクトへ移動するボタンの追加
ゲームを終了するボタンの追加
オプションボタン保留
ボタンスクリプトを流用できるように修正
