+++ ProjectTeamH/Assets/Shimada/Scene/Ending.unity
@@ -404,7 +404,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1312401287}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 910a4f001267d3347bdaa4bd021febbe, type: 3}
+  m_Script: {fileID: 11500000, guid: af3deb81a6181134a9c76fefa69fd333, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!114 &1312401290
@@ -448,12 +448,12 @@ MonoBehaviour:
     m_PersistentCalls:
       m_Calls:
       - m_Target: {fileID: 1312401289}
-        m_MethodName: toGameScene
-        m_Mode: 1
+        m_MethodName: OnClick
+        m_Mode: 3
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
           m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
+          m_IntArgument: 2
           m_FloatArgument: 0
           m_StringArgument: 
           m_BoolArgument: 0
