--- ProjectTeamH/Assets/Shimada/Scripts/EndingButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 910a4f001267d3347bdaa4bd021febbe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
