+++ ProjectTeamH/ProjectSettings/InputManager.asset
@@ -284,11 +284,7 @@ InputManager:
     negativeButton: 
     positiveButton: e
     altNegativeButton: 
-
     altPositiveButton: joystick button 1
-
-    altPositiveButton: joystick button 4
-
     gravity: 1000
     dead: 0.001
     sensitivity: 1000
@@ -362,14 +358,13 @@ InputManager:
     axis: 0
     joyNum: 0
   - serializedVersion: 3
-
     m_Name: Submit
     descriptiveName: 
     descriptiveNegativeName: 
     negativeButton: 
     positiveButton: escape
     altNegativeButton: 
-    altPositiveButton: joystick button 1
+    altPositiveButton: joystick button 0
     gravity: 1000
     dead: 0.001
     sensitivity: 1000
